requirements.txt
bidict==0.22.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
filelock==3.7.1
Flask==2.1.2
Flask-Cors==3.0.10
flask-kafka==0.0.3
Flask-SocketIO==5.2.0
huggingface-hub==0.7.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.1
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
pymongo==4.1.1
pyparsing==3.0.9
python-engineio==4.3.2
python-socketio==5.6.0
PyYAML==6.0
regex==2022.6.2
requests==2.28.0
six==1.16.0
tokenizers==0.12.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.19.4
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
